chore: rename sections to reflect content
