+++ content/sections/1-Story.md
@@ -4,55 +4,55 @@ header: "Our Story So Far"
 subheader: "Catch up as we begin the next chapter!"
 timeline: [
   {
-    imageFileName: "about/1.jpg",
+    imageFileName: "story/1.jpg",
     header: "Sept. 30, 2017",
     subheader: "We first meet",
     content: "In Old City, Philadelphia, Caitlin successfully convinced Justin that she knew anything at all about beer, and impressed him by ordering a pumpkin ale. (She secretly hated it as soon as she tried it.) After dinner, we walked to the other end of the city in search of cocktails, talking for hours and getting to know each other along the way."
   },
   {
-    imageFileName: "about/2.jpg",
+    imageFileName: "story/2.jpg",
     header: "Late 2017 – Early 2018",
     subheader: "A budding romance",
     content: "Together we shared the joys of Philly life (and Justin even convinced Caitlin to spend some time in the ‘burbs).  We began to meet friends and family, and our relationship continued to grow into something we both knew was truly special."
   },
   {
-    imageFileName: "about/3.jpg",
+    imageFileName: "story/3.jpg",
     header: "April 2018 - July 2019",
     subheader: "Long-distance",
     content: "Justin got a promotion and relocated to the Hudson Valley, just outside New York City. For fifteen months, we spent our weekends split between Philly and New York, and came to loathe the Lincoln Tunnel."
   },
   {
-    imageFileName: "about/4.jpg",
+    imageFileName: "story/4.jpg",
     header: "July, 2018",
     subheader: "Our first trip",
     content: "We took a roadtrip up the Blue Ridge Parkway, from Chattanooga, TN to Philadelphia, stopping in Asheville, Roanoke, Charlottesville, Richmond, Williamsburg, and Alexandria along the way. We did some hiking, took in the views, and visited some friends and family."
   },
   {
-    imageFileName: "about/5.jpg",
+    imageFileName: "story/5.jpg",
     header: "August, 2018",
     subheader: "We go international",
     content: "The first international border we crossed together was Canadian (and we absolutely did not begin the trip with an emergency situation involving Caitlin realizing in Buffalo that she had forgotten her passport in Philly). We've been back to visit Toronto and Niagara Falls since then, and have also made it to England and Japan."
   },
   {
-    imageFileName: "about/6.jpg",
+    imageFileName: "story/6.jpg",
     header: "July, 2019",
     subheader: "Reunited",
     content: "After fifteen months of navigating a long-distance relationship, we were finally reunited when Caitlin relocated to Sleepy Hollow, NY. Without the need to travel every weekend, Justin finally had time to nerd out, brewing beer and playing D&D with friends."
   },
   {
-    imageFileName: "about/7.png",
+    imageFileName: "story/7.png",
     header: "July 18, 2020",
     subheader: "Our family grows",
     content: "After relocating to the Maryland suburbs of DC, we adopted the best puppy in the universe, Tanuki, from a shelter in Caitlin's hometown of Waldorf."
   },
     {
-    imageFileName: "about/8.png",
+    imageFileName: "story/8.png",
     header: "October 1, 2020",
     subheader: "We get engaged",
     content: "As we picnicked in Washington Square Park, blocks away from where we first met in Philadelphia, Justin popped the question as Caitlin obvliviously cracked jokes before finally realizing what what going on."
   },
   {
-    imageFileName: "about/9.jpg",
+    imageFileName: "story/9.jpg",
     imageContent: "December\r\n13th\r\n2020",
     header: "December 13, 2020",
     subheader: "The big day!"
