+++ src/components/Charity.scss
@@ -1,5 +1,5 @@
-// Styling for the team section
-.team-member {
+// Styling for the charity section
+.charity {
   margin-bottom: 50px;
   text-align: center;
   .circle {
