--- src/components/ServiceItem.scss
@@ -1,10 +1,0 @@
-// Styling for the services section
-.service-item-image {
-  width: 128px;
-  height: 128px;
-  margin: auto;
-}
-.service-item-heading {
-  margin: 15px 0;
-  text-transform: none;
-}
