+++ src/views/Footer/Footer.jsx
@@ -20,41 +20,39 @@ const Footer = ({ frontmatter }) => {
     <footer className="footer py-3">
       <Container>
         <Row className="align-items-center text-center">
-          {/* <div className="creditWrapper"> */}
-            <Col className="copyright credit">
-              <div className="nameAndLink">
-                ©&nbsp;
-                <a 
-                  href="https://caitlinfloyd.com/"
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  Caitlin Floyd
-                </a>
-                , 2021
-              </div>
-              <div className="socials">
-                {github ? <SocialIcons.Github className="footerSocial" userName={github} /> : null}
-                {linkedin ? <SocialIcons.Linkedin className="footerSocial" userName={linkedin} /> : null}
-              </div>
-            </Col> {/* copyright */}
-            <Col className="anaIsabel credit">
-              <div className="nameAndLink">
-                Photos by&nbsp;
-                <a 
-                  href="https://anaisabelphotography.com/"
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  Ana Isabel Photography
-                </a>
-              </div>
-              <div className="socials">
-                {instagram ? <SocialIcons.Instagram className="footerSocial" userName={instagram} /> : null}
-                {facebook ? <SocialIcons.Facebook className="footerSocial" userName={facebook} /> : null}
-              </div>
-            </Col> {/* anaIsabel */}
-          {/* </div> */}
+          <Col className="copyright credit">
+            <div className="nameAndLink">
+              ©&nbsp;
+              <a 
+                href="https://caitlinfloyd.com/"
+                target="_blank"
+                rel="noreferrer"
+              >
+                Caitlin Floyd
+              </a>
+              , 2021
+            </div>
+            <div className="socials">
+              {github ? <SocialIcons.Github className="footerSocial" userName={github} /> : null}
+              {linkedin ? <SocialIcons.Linkedin className="footerSocial" userName={linkedin} /> : null}
+            </div>
+          </Col> {/* copyright */}
+          <Col className="anaIsabel credit">
+            <div className="nameAndLink">
+              Photos by&nbsp;
+              <a 
+                href="https://anaisabelphotography.com/"
+                target="_blank"
+                rel="noreferrer"
+              >
+                Ana Isabel Photography
+              </a>
+            </div>
+            <div className="socials">
+              {instagram ? <SocialIcons.Instagram className="footerSocial" userName={instagram} /> : null}
+              {facebook ? <SocialIcons.Facebook className="footerSocial" userName={facebook} /> : null}
+            </div>
+          </Col> {/* anaIsabel */}
         </Row>
       </Container>
     </footer>
