+++ src/views/Sections/Ceremony/Ceremony.scss
@@ -1,5 +1,5 @@
-// Styling for the portfolio section
-.portfolio-section {
+// Styling for the ceremony section
+.ceremony-section {
   * {
     z-index: 2;
   }
