--- src/views/Sections/Clients/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Clients";
