--- src/views/Sections/Contact/Contact.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const Contact = () => {
-  return <div />;
-};
-
-Contact.propTypes = {
-};
-
-Contact.defaultProps = {
-};
-
-export default Contact;
