+++ src/views/Sections/Giving/Giving.scss
