--- src/views/Sections/Services/Services.jsx
@@ -1,29 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-import { Row } from "react-bootstrap";
-import SectionHeader from "components/SectionHeader";
-import PageSection from "components/PageSection";
-import "./Services.scss"
-
-const Services = ({ className, frontmatter }) => {
-  if (!frontmatter) {
-    return null;
-  }
-
-  const { anchor, header: rootHeader, subheader: rootSubHeader } = frontmatter;
-
-  return <div/>
-};
-
-Services.propTypes = {
-  className: PropTypes.string,
-  frontmatter: PropTypes.object,
-};
-
-Services.defaultProps = {
-  className: null,
-  frontmatter: null,
-};
-
-export default Services;
