--- src/views/Sections/Services/Services.scss
@@ -1,7 +1,0 @@
-.message {
-    margin: 40px;
-}
-
-@media (min-width: 768px) {
-
-}
