--- src/views/Sections/Services/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Services";
