+++ src/views/Sections/Story/Story.scss
