+++ src/views/Sections/Story/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Story";
