--- src/views/Sections/Team/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Team";
